[Maintenance] Refactor services' ids according to the new convention
