+++ src/Resources/config/admin_routing.yml
@@ -2,10 +2,10 @@ sylius_paypal_plugin_admin_download_payouts_report:
     path: /download/payouts-report/{id}
     methods: [GET]
     defaults:
-        _controller: Sylius\PayPalPlugin\Controller\DownloadPayoutsReportAction
+        _controller: sylius_paypal.controller.download_payouts_report
 
 sylius_paypal_plugin_admin_enable_seller:
     path: /enable-seller/{id}
     methods: [POST]
     defaults:
-        _controller: Sylius\PayPalPlugin\Controller\EnableSellerAction
+        _controller: sylius_paypal.controller.enable_seller
