+++ src/Resources/config/services/listeners/workflow.xml
@@ -1,17 +1,34 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Sylius Sp. z o.o.
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container
+        xmlns="http://symfony.com/schema/dic/services"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
+>
     <services>
         <service id="Sylius\PayPalPlugin\EventListener\Workflow\CompletePayPalOrderListener">
-            <argument type="service" id="Sylius\PayPalPlugin\Processor\PayPalOrderCompleteProcessor" />
+            <argument type="service" id="sylius_paypal.processor.paypal_order_complete" />
 
             <tag name="kernel.event_listener" event="workflow.sylius_order_checkout.completed.complete" priority="100" />
         </service>
+        <service id="sylius_paypal.listener.workflow.complete_paypal_order" alias="Sylius\PayPalPlugin\EventListener\Workflow\CompletePayPalOrderListener" />
 
         <service id="Sylius\PayPalPlugin\EventListener\Workflow\RefundPaymentListener">
-            <argument type="service" id="Sylius\PayPalPlugin\Processor\PaymentRefundProcessorInterface" />
+            <argument type="service" id="sylius_paypal.processor.payment_refund" />
 
             <tag name="kernel.event_listener" event="workflow.sylius_payment.transition.refund" priority="100" />
         </service>
+        <service id="sylius_paypal.listener.workflow.refund_payment" alias="Sylius\PayPalPlugin\EventListener\Workflow\RefundPaymentListener" />
     </services>
 </container>
