+++ src/Resources/config/services/onboarding.xml
@@ -1,6 +1,21 @@
-<?xml version="1.0" encoding="UTF-8" ?>
+<?xml version="1.0" encoding="UTF-8"?>
 
-<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Sylius Sp. z o.o.
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container
+        xmlns="http://symfony.com/schema/dic/services"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
+>
     <services>
         <service
             id="Sylius\PayPalPlugin\Onboarding\Initiator\OnboardingInitiatorInterface"
@@ -10,6 +25,7 @@
             <argument type="service" id="security.helper" />
             <argument>%sylius_paypal.facilitator_url%</argument>
         </service>
+        <service id="sylius_paypal.onboarding.initiator" alias="Sylius\PayPalPlugin\Onboarding\Initiator\OnboardingInitiatorInterface" />
 
         <service
             id="Sylius\PayPalPlugin\Onboarding\Processor\OnboardingProcessorInterface"
@@ -18,7 +34,10 @@
             <argument type="service" id="Http\Discovery\Psr18Client" />
             <argument type="service" id="Sylius\PayPalPlugin\Registrar\SellerWebhookRegistrarInterface" />
             <argument>%sylius_paypal.facilitator_url%</argument>
+            <argument type="service" id="sylius_paypal.http.discovery.psr_18_client" />
+            <argument type="service" id="sylius_paypal.registrar.seller_webhook" />
             <argument type="service" id="Psr\Http\Message\RequestFactoryInterface" />
         </service>
+        <service id="sylius_paypal.onboarding.processor.basic" alias="Sylius\PayPalPlugin\Onboarding\Processor\OnboardingProcessorInterface" />
     </services>
 </container>
