+++ src/Resources/config/services/payum.xml
@@ -1,6 +1,21 @@
-<?xml version="1.0" encoding="UTF-8" ?>
+<?xml version="1.0" encoding="UTF-8"?>
 
-<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Sylius Sp. z o.o.
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container
+        xmlns="http://symfony.com/schema/dic/services"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
+>
     <services>
         <service
             id="sylius.plugin.pay_pal.gateway_factory_builder"
@@ -9,32 +24,37 @@
             <argument>Sylius\PayPalPlugin\Payum\Factory\PayPalGatewayFactory</argument>
             <tag name="payum.gateway_factory_builder" factory="sylius.pay_pal" />
         </service>
+        <service id="sylius_paypal.gateway_factory_builder" alias="sylius.plugin.pay_pal.gateway_factory_builder" />
 
         <service id="Sylius\PayPalPlugin\Payum\Action\AuthorizeAction" public="true">
             <tag name="payum.action" factory="sylius.pay_pal" alias="payum.action.authorize" />
         </service>
+        <service id="sylius_paypal.payum.action.authorize" alias="Sylius\PayPalPlugin\Payum\Action\AuthorizeAction" />
 
         <service id="Sylius\PayPalPlugin\Payum\Action\CaptureAction" public="true">
-            <argument type="service" id="Sylius\PayPalPlugin\Api\CacheAuthorizeClientApiInterface" />
-            <argument type="service" id="Sylius\PayPalPlugin\Api\CreateOrderApiInterface" />
-            <argument type="service" id="Sylius\PayPalPlugin\Provider\UuidProviderInterface" />
+            <argument type="service" id="sylius_paypal.api.cache_authorize_client" />
+            <argument type="service" id="sylius_paypal.api.create_order" />
+            <argument type="service" id="sylius_paypal.provider.uuid" />
             <tag name="payum.action" factory="sylius.pay_pal" alias="payum.action.capture" />
         </service>
+        <service id="sylius_paypal.payum.action.capture" alias="Sylius\PayPalPlugin\Payum\Action\CaptureAction" />
 
         <service id="Sylius\PayPalPlugin\Payum\Action\CompleteOrderAction" public="true">
-            <argument type="service" id="Sylius\PayPalPlugin\Api\CacheAuthorizeClientApiInterface" />
-            <argument type="service" id="Sylius\PayPalPlugin\Api\UpdateOrderApiInterface" />
-            <argument type="service" id="Sylius\PayPalPlugin\Api\CompleteOrderApiInterface" />
-            <argument type="service" id="Sylius\PayPalPlugin\Api\OrderDetailsApiInterface" />
-            <argument type="service" id="Sylius\PayPalPlugin\Processor\PayPalAddressProcessor" />
-            <argument type="service" id="Sylius\PayPalPlugin\Updater\PaymentUpdaterInterface" />
+            <argument type="service" id="sylius_paypal.api.cache_authorize_client" />
+            <argument type="service" id="sylius_paypal.api.update_order" />
+            <argument type="service" id="sylius_paypal.api.complete_order" />
+            <argument type="service" id="sylius_paypal.api.order_details" />
+            <argument type="service" id="sylius_paypal.processor.paypal_address" />
+            <argument type="service" id="sylius_paypal.updater.payment" />
             <argument type="service" id="sylius.state_resolver.order_payment" />
             <argument>null</argument>
             <tag name="payum.action" factory="sylius.pay_pal" alias="payum.action.complete_order" />
         </service>
+        <service id="sylius_paypal.payum.action.complete_order" alias="Sylius\PayPalPlugin\Payum\Action\CompleteOrderAction" />
 
         <service id="Sylius\PayPalPlugin\Payum\Action\ResolveNextRouteAction" public="true">
             <tag name="payum.action" factory="sylius.pay_pal" alias="sylius.resolve_next_route" />
         </service>
+        <service id="sylius_paypal.payum.action.resolve_next_route" alias="Sylius\PayPalPlugin\Payum\Action\ResolveNextRouteAction" />
     </services>
 </container>
